move unit tests according to rust convention
