+++ src/lib.rs
@@ -607,3 +607,52 @@ impl RubiksCube {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_rotation_in_one_direction_then_opposite_should_give_identity() {
+        let my_rubiks_cube = RubiksCube::new();
+        for color in ColorFacet::iterator() {
+            for direction in RotationDirection::iterator() {
+                let rotated_cube = my_rubiks_cube
+                    .rotate(*color, *direction)
+                    .rotate(*color, direction.opposite());
+                assert_eq!(my_rubiks_cube, rotated_cube);
+            }
+        }
+    }
+
+    #[test]
+    fn test_four_same_rotation_should_give_identity() {
+        let my_rubiks_cube = RubiksCube::new();
+        for color in ColorFacet::iterator() {
+            for direction in RotationDirection::iterator() {
+                let rotated_cube = my_rubiks_cube
+                    .rotate(*color, *direction)
+                    .rotate(*color, *direction)
+                    .rotate(*color, *direction)
+                    .rotate(*color, *direction);
+                assert_eq!(my_rubiks_cube, rotated_cube);
+            }
+        }
+    }
+
+    #[test]
+    fn test_one_rotation_should_give_same_as_three_opposite_rotation() {
+        let my_rubiks_cube = RubiksCube::new();
+        for color in ColorFacet::iterator() {
+            for direction in RotationDirection::iterator() {
+                let rotated_cube = my_rubiks_cube.rotate(*color, *direction);
+                let opposite_rotated_cube = my_rubiks_cube
+                    .rotate(*color, direction.opposite())
+                    .rotate(*color, direction.opposite())
+                    .rotate(*color, direction.opposite());
+                assert_eq!(rotated_cube, opposite_rotated_cube);
+            }
+        }
+    }
+
+}
